--- projects/components-demo/src/app/demo-alert/demo-alert.component.scss
@@ -1,8 +1,0 @@
-dp-button {
-  display: inline-block;
-  margin-top: 1rem;
-}
-
-td:first-child {
-  font-family: Consolas, "courier new";
-}
