--- projects/components-demo/src/app/demo-checkbox/demo-checkbox.component.scss
@@ -1,11 +1,0 @@
-dp-button {
-  margin-right: 0.5rem;
-}
-
-dp-checkbox {
-  display: block;
-}
-
-td:first-child {
-  font-family: Consolas, "courier new";
-}
