--- projects/components-demo/src/app/demo-date-picker/demo-date-picker.component.html
@@ -1,21 +1,0 @@
-<div class="demo-wrapper">
-  <div class="demo-title">Date Picker</div>
-
-  <div class="demo-border">
-    <label class="demo-subtitle">Date</label>
-    <dp-date-picker placeholder="YYYY-MM-DD"
-                    minDate="2015-01-01"
-                    [formControl]="formControl"></dp-date-picker>
-    <div class="demo-errors" *ngIf="errors">{{errors}}</div>
-
-    <div style="margin-top: 1rem; margin-bottom: 1rem;">
-      <dp-button color="primary" *ngIf="this.formControl.disabled" (click)="enable()">Enable</dp-button>
-      <dp-button color="primary" *ngIf="!this.formControl.disabled" (click)="disable()">Disable</dp-button>
-      <dp-button color="secondary" (click)="error()">Error</dp-button>
-    </div>
-
-    <div class="demo-value">
-      Value: {{ formControl.value }}
-    </div>
-  </div>
-</div>
