--- projects/components-demo/src/app/demo-date-picker/demo-date-picker.component.ts
@@ -1,27 +1,0 @@
-import { Component } from '@angular/core';
-import { FormControl } from '@angular/forms';
-
-@Component({
-  selector: 'dp-demo-date-picker',
-  templateUrl: './demo-date-picker.component.html',
-  styleUrls: ['./demo-date-picker.component.scss']
-})
-export class DemoDatePickerComponent {
-  formControl = new FormControl('2016-03-03');
-
-  enable(): void {
-    this.formControl.enable();
-  }
-
-  disable(): void {
-    this.formControl.disable();
-  }
-
-  error(): void {
-    this.formControl.setErrors({somekey: 'this is an error'});
-  }
-
-  get errors(): string[] {
-    return this.formControl.errors ? Object.values(this.formControl.errors) : null;
-  }
-}
