+++ projects/components-demo/src/app/demos/demo-alert/demo-alert.component.scss
@@ -1,0 +1,4 @@
+dp-button {
+  display: inline-block;
+  margin-top: 1rem;
+}
