+++ projects/components-demo/src/app/demos/demo-checkbox/demo-checkbox.component.scss
@@ -1,3 +1,7 @@
 dp-button {
   margin-right: 0.5rem;
 }
+
+dp-checkbox {
+  display: block;
+}
