+++ projects/components-demo/src/app/demos/demo-chip/demo-chip.component.ts
