+++ projects/components-demo/src/app/demos/demo-date-picker/demo-date-picker.component.html
@@ -1,0 +1,117 @@
+<div class="component-intro">
+  <h1>Date Picker</h1>
+  <p>
+    A text input component for inputting dates that includes a calendar popper for visual selection.
+    Supports <code>ControlValueAccessor</code>.
+  </p>
+</div>
+
+<div class="demo-wrapper">
+  <h4>Date:</h4>
+  <dp-date-picker placeholder="YYYY-MM-DD"
+                  minDate="2015-01-01"
+                  [formControl]="formControl"></dp-date-picker>
+  <div class="demo-errors" *ngIf="errors">{{ errors }}</div>
+
+  <div class="demo-btn-group">
+    <dp-button color="primary" *ngIf="this.formControl.disabled" (click)="enable()">Enable</dp-button>
+    <dp-button color="primary" *ngIf="!this.formControl.disabled" (click)="disable()">Disable</dp-button>
+    <dp-button color="secondary" (click)="error()">Error</dp-button>
+  </div>
+
+  <div class="demo-value">Value: {{ formControl.value }}</div>
+</div>
+
+<div class="component-info">
+  <dp-tab-group>
+    <dp-tab label="Documentation">
+      <h3>Getting Started</h3>
+      <p>The date picker component can be used with Angular Form Control.</p>
+      <app-codeblock>{{ codeGetStarted }}</app-codeblock>
+      <h4>Import</h4>
+      <app-codeblock>Currently everything is imported by DevproComponentsModule.</app-codeblock>
+
+      <h3>Minimum & Maximum Date Range</h3>
+      <p>A minimum and/or maximum date can be set to limit which dates can be input into the field to prevent things
+        like scheduling a date in the past.</p>
+      <app-codeblock>{{ codeMinMax }}</app-codeblock>
+
+      <h3>Properties</h3>
+      <table>
+        <tr>
+          <th>Name</th>
+          <th>Type</th>
+          <th>Default Value</th>
+          <th>Description</th>
+        </tr>
+        <tr>
+          <td>placeholder</td>
+          <td>string</td>
+          <td>null</td>
+          <td>Placeholder for the input field.</td>
+        </tr>
+        <tr>
+          <td>value</td>
+          <td>any</td>
+          <td>null</td>
+          <td>Value of the date picker.</td>
+        </tr>
+        <tr>
+          <td>disabled</td>
+          <td>boolean</td>
+          <td>false</td>
+          <td>Disabled attribute of the date picker, when true it will visibly change the input field and make it un-clickable.</td>
+        </tr>
+        <tr>
+          <td>minDate</td>
+          <td>-</td>
+          <td>-</td>
+          <td>Minimum date allowed to be input/selected.</td>
+        </tr>
+        <tr>
+          <td>maxDate</td>
+          <td>-</td>
+          <td>-</td>
+          <td>Maximum date allowed to be input/selected.</td>
+        </tr>
+      </table>
+
+      <h3>Events</h3>
+      <table>
+        <tr>
+          <th>Name</th>
+          <th>Parameters</th>
+          <th>Description</th>
+        </tr>
+        <tr>
+          <td>notifyChanges</td>
+          <td>change: Emitted string that is a new change in value.</td>
+          <td>Notifies changes to the state of the field and its value.</td>
+        </tr>
+      </table>
+
+      <h3>Styling</h3>
+      <table>
+        <tr>
+          <th>Name</th>
+          <th>Element</th>
+        </tr>
+        <tr>
+          <td>input-wrapper</td>
+          <td>Outermost element containing the input and calendar popper.</td>
+        </tr>
+      </table>
+
+      <h3>Methods</h3>
+      <p>None.</p>
+
+      <h3>Dependencies</h3>
+      <p><a href="https://moment.github.io/luxon/index.html">Luxon</a> DateTime</p>
+    </dp-tab>
+
+    <dp-tab label="Source Code">
+      <app-codeblock>{{ codeHTML }}</app-codeblock>
+      <app-codeblock>{{ codeTS }}</app-codeblock>
+    </dp-tab>
+  </dp-tab-group>
+</div>
