+++ projects/components-demo/src/app/demos/demo-date-picker/demo-date-picker.component.scss
@@ -1,0 +1,12 @@
+dp-button {
+  margin-right: 0.5rem;
+}
+
+dp-date-picker {
+  width: 75%;
+}
+
+a {
+  color: rgba(59, 130, 246, 1);
+  text-decoration: none;
+}
